+++ authorize/model_entity_array__embedded_permissions_inner.go
@@ -40,23 +40,6 @@ func ApplicationRolePermissionAsEntityArrayEmbeddedPermissionsInner(v *Applicati
 func (dst *EntityArrayEmbeddedPermissionsInner) UnmarshalJSON(data []byte) error {
 
 	var err error
-	// try to unmarshal JSON data into ApplicationResourcePermission
-	err = json.Unmarshal(data, &dst.ApplicationResourcePermission)
-	if err == nil {
-		jsonApplicationResourcePermission, _ := json.Marshal(dst.ApplicationResourcePermission)
-		if string(jsonApplicationResourcePermission) == "{}" { // empty struct
-			dst.ApplicationResourcePermission = nil
-		} else {
-			if dst.ApplicationResourcePermission.Action == "" { // we expect an action for this data type
-				dst.ApplicationResourcePermission = nil
-			} else {
-				return nil // data stored in dst.ApplicationResourcePermission, return on the first match
-			}
-		}
-	} else {
-		dst.ApplicationResourcePermission = nil
-	}
-
 	// try to unmarshal JSON data into ApplicationRolePermission
 	err = json.Unmarshal(data, &dst.ApplicationRolePermission)
 	if err == nil {
@@ -64,7 +47,7 @@ func (dst *EntityArrayEmbeddedPermissionsInner) UnmarshalJSON(data []byte) error
 		if string(jsonApplicationRolePermission) == "{}" { // empty struct
 			dst.ApplicationRolePermission = nil
 		} else {
-			if dst.ApplicationRolePermission.Permission != nil {
+			if dst.ApplicationRolePermission.Key != nil {
 				return nil // data stored in dst.ApplicationRolePermission, return on the first match
 			} else {
 				dst.ApplicationRolePermission = nil
@@ -72,6 +55,23 @@ func (dst *EntityArrayEmbeddedPermissionsInner) UnmarshalJSON(data []byte) error
 		}
 	} else {
 		dst.ApplicationRolePermission = nil
+	}
+
+	// try to unmarshal JSON data into ApplicationResourcePermission
+	err = json.Unmarshal(data, &dst.ApplicationResourcePermission)
+	if err == nil {
+		jsonApplicationResourcePermission, _ := json.Marshal(dst.ApplicationResourcePermission)
+		if string(jsonApplicationResourcePermission) == "{}" { // empty struct
+			dst.ApplicationResourcePermission = nil
+		} else {
+			if dst.ApplicationResourcePermission.Action != "" { // we expect an resource for this data type
+				dst.ApplicationResourcePermission = nil
+			} else {
+				return nil // data stored in dst.ApplicationResourcePermission, return on the first match
+			}
+		}
+	} else {
+		dst.ApplicationResourcePermission = nil
 	}
 
 	return fmt.Errorf("Data failed to match schemas in anyOf(EntityArrayEmbeddedPermissionsInner)")