+++ pingone/model/region.go
@@ -9,6 +9,7 @@ import (
 )
 
 type RegionMapping struct {
+	// Deprecated: Use APICode instead
 	Region    string
 	URLSuffix string
 	APICode   management.EnumRegionCode
@@ -36,6 +37,11 @@ func init() {
 			APICode:   management.ENUMREGIONCODE_AP,
 		},
 		{
+			Region:    "Australia-AsiaPacific",
+			URLSuffix: "com.au",
+			APICode:   management.ENUMREGIONCODE_AU,
+		},
+		{
 			Region:    "Canada",
 			URLSuffix: "ca",
 			APICode:   management.ENUMREGIONCODE_CA,
@@ -43,6 +49,7 @@ func init() {
 	}
 }
 
+// Deprecated: Use FindRegionByAPICode instead
 func FindRegionByName(region string) RegionMapping {
 
 	idx := slices.IndexFunc(regionMappingList, func(c RegionMapping) bool { return c.Region == region })