+++ verify/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.5.0 (Unreleased)
+
+* **Enhancement** Add the `com.au` top level domain to the connection configuration. [#351](https://github.com/patrickcping/pingone-go-sdk-v2/pull/351)
+
 # v0.4.1 (2024-01-12)
 
 * **Note** Upgrade GO to `v1.21` [#306](https://github.com/patrickcping/pingone-go-sdk-v2/pull/306)